Change from jdbc to jpa connection

+++ src/main/kotlin/no/stunor/origo/batch/data/OrganisationRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface OrganisationRepository : CrudRepository<Organisation, Long> {
+interface OrganisationRepository : CrudRepository<Organisation, Int> {
     fun findAllByEventorId(eventorId: String): List<Organisation>
 }
